feat: remove baseDeployCommand.ts
