+++ packages/salesforcedx-vscode-core/src/commands/baseDeployRetrieve.ts
@@ -30,14 +30,13 @@ import * as vscode from 'vscode';
 import { channelService, OUTPUT_CHANNEL } from '../channels';
 import { PersistentStorageService } from '../conflict/persistentStorageService';
 import { TELEMETRY_METADATA_COUNT } from '../constants';
-import { WorkspaceContext } from '../context';
-import { workspaceContextUtils } from '../context';
+import { WorkspaceContext, workspaceContextUtils } from '../context';
 import { handleDeployDiagnostics } from '../diagnostics';
 import { nls } from '../messages';
 import { setApiVersionOn } from '../services/sdr/componentSetUtils';
 import { DeployQueue } from '../settings';
 import { SfdxPackageDirectories } from '../sfdxProject';
-import { BaseDeployExecutor } from './baseDeployCommand';
+import { ForceSourcePushExecutor } from './forceSourcePush';
 import { createComponentCount, formatException } from './util';
 
 type DeployRetrieveResult = DeployResult | RetrieveResult;
@@ -138,7 +137,7 @@ export abstract class DeployExecutor<T> extends DeployRetrieveExecutor<T> {
   ): Promise<void> {
     try {
       if (result) {
-        BaseDeployExecutor.errorCollection.clear();
+        ForceSourcePushExecutor.errorCollection.clear();
 
         // Update Persistent Storage for the files that were deployed
         PersistentStorageService.getInstance().setPropertiesForFilesDeploy(
@@ -151,7 +150,10 @@ export abstract class DeployExecutor<T> extends DeployRetrieveExecutor<T> {
 
         const success = result.response.status === RequestStatus.Succeeded;
         if (!success) {
-          handleDeployDiagnostics(result, BaseDeployExecutor.errorCollection);
+          handleDeployDiagnostics(
+            result,
+            ForceSourcePushExecutor.errorCollection
+          );
         }
       }
     } finally {
