+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/forceSourceDeploy.test.ts
@@ -16,10 +16,7 @@ import { Row, Table } from '@salesforce/salesforcedx-utils-vscode';
 import { expect } from 'chai';
 import { stub } from 'sinon';
 import { channelService } from '../../../src/channels';
-import {
-  BaseDeployExecutor,
-  ForceSourcePushExecutor
-} from '../../../src/commands';
+import { ForceSourcePushExecutor } from '../../../src/commands';
 import { nls } from '../../../src/messages';
 
 describe('Correctly output deploy results', () => {
